backend/requirements.txt
Django==3.2
djangorestframework==3.12.0
django-filter==22.1
djoser==2.1.0
requests==2.28.1
python-dotenv==0.20.0
psycopg2-binary
pytest==7.1.2
pillow==9.2.0
